pyproject.toml
[tool.poetry]
name = "thebro"
version = "0.1.0"
description = ""
authors = ["Bryce Jenkins <bryce@durish.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
hikari = "^2.1.0"
hikari-lightbulb = "^2.3.5.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
